LayoutTests/fast/scrolling/mac/stateless-user-scroll-scrollend-expected.txt

PASS windowScrollendEventCount == 1 is true
PASS successfullyParsed is true

TEST COMPLETE
